organize code by making utils files
